+++ buildSrc/src/main/kotlin/BuildConfig.kt
@@ -3,7 +3,7 @@ package com.unciv.build
 
 object BuildConfig {
     const val appName = "UncivCN"
-    const val appCodeNumber = 1256
-    const val appVersion = "4.21.10.1"
+    const val appCodeNumber = 1257
+    const val appVersion = "4.21.10.2"
     const val identifier = "com.unciv.app"
 }